refactoring architechture
