+++ applications.csv
@@ -1,38 +1,47 @@
-"id","sum","credit","maturity","rate","date","status","clientID"
-"1482066374535","35000.0","0","20","10.0","1482066374535","0","0"
-"1482066374543","105000.0","0","20","10.0","1482066374543","0","0"
-"1482066374549","140000.0","0","20","10.0","1482066374549","0","0"
-"1482066374557","175000.0","0","20","10.0","1482066374557","0","0"
-"1482066374568","210000.0","0","20","10.0","1482066374568","0","0"
-"1482066374582","245000.0","0","20","10.0","1482066374582","0","0"
-"1482066374593","280000.0","0","20","10.0","1482066374593","0","0"
-"1482066374606","315000.0","0","20","10.0","1482066374606","0","0"
-"1482066374539","7000000.0","0","20","10.0","1482066374539","0","0"
-"1482066383969","35000.0","0","20","10.0","1482066383969","0","0"
-"1482066383998","105000.0","0","20","10.0","1482066383998","0","0"
-"1482066384014","140000.0","0","20","10.0","1482066384014","0","0"
-"1482066384034","175000.0","0","20","10.0","1482066384034","0","0"
-"1482066384053","210000.0","0","20","10.0","1482066384053","0","0"
-"1482066384072","245000.0","0","20","10.0","1482066384072","0","0"
-"1482066384093","280000.0","0","20","10.0","1482066384093","0","0"
-"1482066384115","315000.0","0","20","10.0","1482066384115","0","0"
-"1482066383983","7000000.0","0","20","10.0","1482066383983","0","0"
-"1482069941234","35000.0","0","20","10.0","1482069941234","0","0"
-"1482069941284","105000.0","0","20","10.0","1482069941284","0","0"
-"1482069941310","140000.0","0","20","10.0","1482069941310","0","0"
-"1482069941338","175000.0","0","20","10.0","1482069941338","0","0"
-"1482069941368","210000.0","0","20","10.0","1482069941368","0","0"
-"1482069941397","245000.0","0","20","10.0","1482069941397","0","0"
-"1482069941430","280000.0","0","20","10.0","1482069941430","0","0"
-"1482069941463","315000.0","0","20","10.0","1482069941463","0","0"
-"1482069941258","7000000.0","0","20","10.0","1482069941258","0","0"
-"1482077945604","0.0","0","20","10.0","1482077945604","0","0"
-"1482077945637","35000.0","0","20","10.0","1482077945637","0","0"
-"1482077945671","70000.0","0","20","10.0","1482077945671","0","0"
-"1482077945706","105000.0","0","20","10.0","1482077945706","0","0"
-"1482077945743","140000.0","0","20","10.0","1482077945743","0","0"
-"1482077945781","175000.0","0","20","10.0","1482077945781","0","0"
-"1482077945820","210000.0","0","20","10.0","1482077945820","0","0"
-"1482077945861","245000.0","0","20","10.0","1482077945861","0","0"
-"1482077945903","280000.0","0","20","10.0","1482077945903","0","0"
-"1482077945945","315000.0","0","20","10.0","1482077945945","0","0"
+"id","sum","credit","maturity","rate","date","status","clientID","typeCredit"
+"1482066374535","35000.0","0","20","10.0","1482066374535","0","0","0"
+"1482066374543","105000.0","0","20","10.0","1482066374543","0","0","0"
+"1482066374549","140000.0","0","20","10.0","1482066374549","0","0","0"
+"1482066374557","175000.0","0","20","10.0","1482066374557","0","0","0"
+"1482066374568","210000.0","0","20","10.0","1482066374568","0","0","0"
+"1482066374582","245000.0","0","20","10.0","1482066374582","0","0","0"
+"1482066374593","280000.0","0","20","10.0","1482066374593","0","0","0"
+"1482066374606","315000.0","0","20","10.0","1482066374606","0","0","0"
+"1482066374539","7000000.0","0","20","10.0","1482066374539","0","0","0"
+"1482066383969","35000.0","0","20","10.0","1482066383969","0","0","0"
+"1482066383998","105000.0","0","20","10.0","1482066383998","0","0","0"
+"1482066384014","140000.0","0","20","10.0","1482066384014","0","0","0"
+"1482066384034","175000.0","0","20","10.0","1482066384034","0","0","0"
+"1482066384053","210000.0","0","20","10.0","1482066384053","0","0","0"
+"1482066384072","245000.0","0","20","10.0","1482066384072","0","0","0"
+"1482066384093","280000.0","0","20","10.0","1482066384093","0","0","0"
+"1482066384115","315000.0","0","20","10.0","1482066384115","0","0","0"
+"1482066383983","7000000.0","0","20","10.0","1482066383983","0","0","0"
+"1482069941234","35000.0","0","20","10.0","1482069941234","0","0","0"
+"1482069941284","105000.0","0","20","10.0","1482069941284","0","0","0"
+"1482069941310","140000.0","0","20","10.0","1482069941310","0","0","0"
+"1482069941338","175000.0","0","20","10.0","1482069941338","0","0","0"
+"1482069941368","210000.0","0","20","10.0","1482069941368","0","0","0"
+"1482069941397","245000.0","0","20","10.0","1482069941397","0","0","0"
+"1482069941430","280000.0","0","20","10.0","1482069941430","0","0","0"
+"1482069941463","315000.0","0","20","10.0","1482069941463","0","0","0"
+"1482069941258","7000000.0","0","20","10.0","1482069941258","0","0","0"
+"1482077945637","35000.0","0","20","10.0","1482077945637","0","0","0"
+"1482077945706","105000.0","0","20","10.0","1482077945706","0","0","0"
+"1482077945743","140000.0","0","20","10.0","1482077945743","0","0","0"
+"1482077945781","175000.0","0","20","10.0","1482077945781","0","0","0"
+"1482077945820","210000.0","0","20","10.0","1482077945820","0","0","0"
+"1482077945861","245000.0","0","20","10.0","1482077945861","0","0","0"
+"1482077945903","280000.0","0","20","10.0","1482077945903","0","0","0"
+"1482077945945","315000.0","0","20","10.0","1482077945945","0","0","0"
+"1482077945671","7000000.0","0","20","10.0","1482077945671","0","0","0"
+"1482098168485","0.0","0","20","10.0","1482098168485","0","0","0"
+"1482098168528","35000.0","0","20","10.0","1482098168528","0","0","0"
+"1482098168573","70000.0","0","20","10.0","1482098168573","0","0","0"
+"1482098168620","105000.0","0","20","10.0","1482098168620","0","0","0"
+"1482098168666","140000.0","0","20","10.0","1482098168666","0","0","0"
+"1482098168714","175000.0","0","20","10.0","1482098168714","0","0","0"
+"1482098168763","210000.0","0","20","10.0","1482098168763","0","0","0"
+"1482098168820","245000.0","0","20","10.0","1482098168820","0","0","0"
+"1482098168874","280000.0","0","20","10.0","1482098168874","0","0","0"
+"1482098168928","315000.0","0","20","10.0","1482098168928","0","0","0"
