+++ clients.csv
@@ -26,7 +26,6 @@
 "1482069940829","7","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
 "1482069940863","8","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
 "1482069940897","9","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
-"1482077944792","0","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
 "1482077944830","1","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
 "1482077944868","2","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
 "1482077944905","3","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
@@ -36,3 +35,13 @@
 "1482077945068","7","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
 "1482077945112","8","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
 "1482077945158","9","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
+"1482098167949","0","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
+"1482098168003","1","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
+"1482098168050","2","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
+"1482098168098","3","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
+"1482098168149","4","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
+"1482098168200","5","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
+"1482098168252","6","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
+"1482098168316","7","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
+"1482098168372","8","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
+"1482098168428","9","0","0","0.0","0.0","0.0","0.0","0.0","true","true","true","Name","Work","Street","234567","276543","pd","null"
