--- credits.csv
@@ -1,38 +1,0 @@
-"id","status","date","typeCredit"
-"1482066374445","1","0","0"
-"1482066374460","3","0","0"
-"1482066374466","4","0","0"
-"1482066374473","5","0","0"
-"1482066374483","6","0","0"
-"1482066374494","7","0","0"
-"1482066374505","8","0","0"
-"1482066374516","9","0","0"
-"1482066374456","200","0","0"
-"1482066383792","1","0","0"
-"1482066383820","3","0","0"
-"1482066383836","4","0","0"
-"1482066383853","5","0","0"
-"1482066383871","6","0","0"
-"1482066383891","7","0","0"
-"1482066383911","8","0","0"
-"1482066383934","9","0","0"
-"1482066383805","200","0","0"
-"1482069940954","1","0","0"
-"1482069941002","3","0","0"
-"1482069941029","4","0","0"
-"1482069941056","5","0","0"
-"1482069941084","6","0","0"
-"1482069941114","7","0","0"
-"1482069941145","8","0","0"
-"1482069941177","9","0","0"
-"1482069940978","200","0","0"
-"1482077945205","0","0","0"
-"1482077945238","1","0","0"
-"1482077945273","2","0","0"
-"1482077945309","3","0","0"
-"1482077945345","4","0","0"
-"1482077945383","5","0","0"
-"1482077945424","6","0","0"
-"1482077945470","7","0","0"
-"1482077945517","8","0","0"
-"1482077945561","9","0","0"
