+++ delay.csv
@@ -26,9 +26,7 @@
 "1482069940324","8","8","8.0"
 "1482069940327","9","9","9.0"
 "1482069940307","2","2","200.0"
-"1482077944370","0","0","0.0"
 "1482077944374","1","1","1.0"
-"1482077944378","2","2","2.0"
 "1482077944383","3","3","3.0"
 "1482077944386","4","4","4.0"
 "1482077944389","5","5","5.0"
@@ -36,3 +34,14 @@
 "1482077944396","7","7","7.0"
 "1482077944399","8","8","8.0"
 "1482077944402","9","9","9.0"
+"1482077944378","2","2","200.0"
+"1482098167421","0","0","0.0"
+"1482098167429","1","1","1.0"
+"1482098167432","2","2","2.0"
+"1482098167436","3","3","3.0"
+"1482098167439","4","4","4.0"
+"1482098167443","5","5","5.0"
+"1482098167447","6","6","6.0"
+"1482098167450","7","7","7.0"
+"1482098167456","8","8","8.0"
+"1482098167459","9","9","9.0"
