+++ payments.csv
@@ -26,9 +26,7 @@
 "1482069940548","1482069940548","8"
 "1482069940579","1482069940579","9"
 "1482069940383","1482069940383","200"
-"1482077944408","1482077944408","0"
 "1482077944440","1482077944440","1"
-"1482077944474","1482077944474","2"
 "1482077944510","1482077944510","3"
 "1482077944546","1482077944546","4"
 "1482077944585","1482077944585","5"
@@ -36,3 +34,14 @@
 "1482077944664","1482077944664","7"
 "1482077944705","1482077944705","8"
 "1482077944747","1482077944747","9"
+"1482077944474","1482077944474","200"
+"1482098167464","1482098167464","0"
+"1482098167508","1482098167508","1"
+"1482098167553","1482098167553","2"
+"1482098167599","1482098167599","3"
+"1482098167646","1482098167646","4"
+"1482098167693","1482098167693","5"
+"1482098167742","1482098167742","6"
+"1482098167792","1482098167792","7"
+"1482098167845","1482098167845","8"
+"1482098167896","1482098167896","9"
