+++ src/main/java/ru/sfedu/mydiplom/model/dto/Payments.java
@@ -3,8 +3,6 @@ package ru.sfedu.mydiplom.model.dto;
 import java.io.Serializable;
 
 public class Payments extends GenericDto{
-
-    private long id = 1l;
     private long date;
     private long credit;
     
@@ -15,7 +13,6 @@ public class Payments extends GenericDto{
 
     public Payments(long id, long date, long credit) throws InterruptedException {
         super(ClassType.PMT, id);
-        this.id = id;
         this.date = date;
         this.credit = credit;
         Thread.sleep(1);
@@ -35,16 +32,6 @@ public class Payments extends GenericDto{
 
     public void setDate(long date) {
         this.date = date;
-    }
-
-    @Override
-    public long getId() {
-        return id;
-    }
-
-    @Override
-    public void setId(long id) {
-        this.id = id;
     }
 
     @Override
