+++ typecredits.csv
@@ -26,9 +26,7 @@
 "1482069941716","Name","8","8","8.0","8.0","8.0","8.0"
 "1482069941748","Name","9","9","9.0","9.0","9.0","9.0"
 "1482069941544","Name","2","2","200.0","2.0","2.0","2.0"
-"1482077945991","Name","0","0","0.0","0.0","0.0","0.0"
 "1482077946026","Name","1","1","1.0","1.0","1.0","1.0"
-"1482077946061","Name","2","2","2.0","2.0","2.0","2.0"
 "1482077946097","Name","3","3","3.0","3.0","3.0","3.0"
 "1482077946134","Name","4","4","4.0","4.0","4.0","4.0"
 "1482077946173","Name","5","5","5.0","5.0","5.0","5.0"
@@ -36,3 +34,14 @@
 "1482077946255","Name","7","7","7.0","7.0","7.0","7.0"
 "1482077946298","Name","8","8","8.0","8.0","8.0","8.0"
 "1482077946340","Name","9","9","9.0","9.0","9.0","9.0"
+"1482077946061","Name","2","2","200.0","2.0","2.0","2.0"
+"1482098168982","Name","0","0","0.0","0.0","0.0","0.0"
+"1482098169027","Name","1","1","1.0","1.0","1.0","1.0"
+"1482098169072","Name","2","2","2.0","2.0","2.0","2.0"
+"1482098169119","Name","3","3","3.0","3.0","3.0","3.0"
+"1482098169168","Name","4","4","4.0","4.0","4.0","4.0"
+"1482098169217","Name","5","5","5.0","5.0","5.0","5.0"
+"1482098169269","Name","6","6","6.0","6.0","6.0","6.0"
+"1482098169320","Name","7","7","7.0","7.0","7.0","7.0"
+"1482098169373","Name","8","8","8.0","8.0","8.0","8.0"
+"1482098169426","Name","9","9","9.0","9.0","9.0","9.0"
